# docs/ai-openapi.yaml
openapi: 3.1.0
info:
  title: SysadminAnywhere REST API
  version: 5.0.0
  description: >-
    Examples for the implemented Directory, Incident, and Inventory service APIs.
    All operations except authentication require an administrator JWT obtained from
    POST /api/ldap/authenticate. See api-reference-ru.md and api-reference-en.md
    for service URLs and usage notes.
servers:
  - url: http://localhost:8081
    description: Directory service (local development)
  - url: http://localhost:8082
    description: Inventory service (local development)
  - url: http://localhost:8083
    description: Incident service (local development)
security:
  - bearerAuth: []
paths:
...
  /api/ldap/authenticate:
    post:
      servers:
        - url: http://localhost:8081
      tags: [Authentication]
      summary: Authenticate against Active Directory
      operationId: authenticate
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              username: administrator
              password: change-me
              service: sysadminanywhere
      responses:
        '200':
          description: JWT and authenticated account details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwtResponse'
              example:
                token: eyJhbGciOiJIUzI1NiJ9.example.signature
                username: administrator
                roles: [ROLE_ADMIN]
        '401':
          description: Invalid username or password
  /api/ldap/search:
    post:
      servers:
        - url: http://localhost:8081
      tags: [Directory]
      summary: Search Active Directory with an LDAP filter
      operationId: searchDirectory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
            example:
              distinguishedName: DC=example,DC=com
              filter: (objectClass=user)
              searchScope: 2
              attributes: [cn, mail, sAMAccountName]
      responses:
        '200':
          description: Matching directory entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DirectoryEntry'
        '400':
          description: Invalid DN, LDAP filter, or search scope
...
  /api/ldap/members/bulk:
    post:
      servers:
        - url: http://localhost:8081
      tags: [Groups, Bulk operations]
      summary: Add or remove multiple objects from a group
      operationId: bulkChangeGroupMembers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkGroupMembershipRequest'
            example:
              memberDistinguishedNames:
                - CN=Alex Morgan,OU=Users,DC=example,DC=com
                - CN=WS-042,OU=Computers,DC=example,DC=com
              groupDistinguishedName: CN=Helpdesk,OU=Groups,DC=example,DC=com
              remove: false
      responses:
        '200':
          description: Per-object operation results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkOperationResult'
        '400':
          description: Invalid group or member distinguished name
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string, example: administrator }
        password: { type: string, format: password }
        service: { type: string, example: sysadminanywhere }
    JwtResponse:
      type: object
      properties:
        token: { type: string }
        username: { type: string }
        roles:
          type: array
          items: { type: string }
    SearchRequest:
      type: object
      required: [filter, searchScope]
      properties:
        distinguishedName: { type: string }
        filter: { type: string, example: (objectClass=user) }
        searchScope:
          type: integer
          description: '0 = base object, 1 = one level, 2 = subtree.'
          enum: [0, 1, 2]
        attributes:
          type: array
          items: { type: string }
...
    BulkGroupMembershipRequest:
      type: object
      required: [memberDistinguishedNames, groupDistinguishedName]
      properties:
        memberDistinguishedNames:
          type: array
          minItems: 1
          items: { type: string }
        groupDistinguishedName: { type: string }
        remove: { type: boolean, default: false }
    DirectoryEntry:
      type: object
      description: Directory attributes are returned according to the requested attributes and object type.
      additionalProperties: true
      properties:
        distinguishedName: { type: string }
        cn: { type: string }
        displayName: { type: string }
        samAccountName: { type: string }
        emailAddress: { type: string }
...
    BulkOperationResult:
      type: object
      description: Per-object success and failure details returned by bulk operations.
      properties:
        updated: { type: integer }
        failures:
          type: array
          items: { type: string }
...